Updated the documentation
Updated the documentation for ID3 (creating random trees), Adaboost, and Bagging.

Also hopefully fixed a bug for maxatrs in ID3 (where it could attempt to sample more items than there were in the attributes list), though the fact that this bug hasn't actually appeared in practice yet may be a red flag.